Geographic data visualisation and Map Generation
This project is one of academic projects given to us in Geographic Information System (GIS) Course.

Created by: Pranav Pandya (Me) and 
Kartikey Hadiya

We sampled information for pollution emmision in Delhi, India. 

Pollution data was obtained from: https://data.gov.in/resources/real-time-air-quality-index-various-locations

Pollution index data can be obtained from: https://cpcb.nic.in/RealTimeAirQualityData.php

Pollution data only had address of Indian Meteorological Department, so each station was located in Google Earth and pin points were added at each station.

Then in the side bar containing those pins on right-click, a new folder was added and all the pins were added in that new folder in google earth. Then that folder was saved as kml file.

This kml file was uploaded to Mygeodata: https://mygeodata.cloud/converter/kml-to-csv and was converted into csv.

Then the csv file was opened and coordinates were copied in pollution data file. That file was later saved as csv and imported in ArcGIS and xy data was displayed.

Shapefile was obtained from web search, which is attached as well. That shapefile was imported in ArcGIS and the final view was generated which is shown in the picture.